analytics/ml_predictor.py
"""Machine learning predictor and simulator for Solana MEV opportunities.
This script performs conservative profit estimation using historical liquidity, prices,
and simulated transaction outcomes. Replace placeholders with real RPC calls.
"""

import yaml
import time
import random
from typing import Dict

# Load global configuration
with open('../config.yaml', 'r') as f:
    CONFIG = yaml.safe_load(f)

def simulate_transaction(tx_payload: Dict) -> Dict:
    """Simulate a transaction payload and return a dictionary with
    'success', 'profit_usd', and 'compute_units' as example fields.
    This function is a placeholder for RPC-based simulateTransaction calls.
    """
    # TODO: replace with real simulation via Solana RPC simulateTransaction
    # Simulate a random outcome for demonstration purposes
    return {
        'success': random.choice([True, True, False]),
        'profit_usd': round(random.uniform(-1.0, 15.0), 6),
        'compute_units': random.randint(1000, 200000)
    }

def evaluate_opportunity(opportunity: Dict) -> float:
    """Run ML models or simple heuristics to estimate expected profit.
    Returns expected profit in USD.
    """
    # Placeholder heuristic
    sim = simulate_transaction(opportunity)
    if not sim['success']:
        return -0.1
    return sim['profit_usd'] - 0.01 * sim['compute_units'] / 100000.0

if __name__ == '__main__':
    print("Starting ML predictor (placeholder). Press Ctrl+C to exit.")
    try:
        while True:
            # Example opportunity payload (replace with real data from mempool)
            example = {'type': 'arbitrage', 'path': ['raydium','orca'], 'volume': 100}
            expected = evaluate_opportunity(example)
            print(f"Estimated profit: ${expected:.6f}")
            time.sleep(5)
    except KeyboardInterrupt:
        print("Exiting ML predictor.")
