analytics/requirements.txt
numpy
pandas
scikit-learn
pyyaml
requests
